pyproject.toml
[tool.poetry]
name = "pp-photo-placer"
version = "0.1.0"
description = ""
authors = ["Josef Kolář <thejoeejoee@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rich-click = "^1.2.1"
fpdf = "^1.7.2"
imagesize = "^1.3.0"
Pillow = "^9.0.1"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
